requirements.txt
# Core dependencies
click>=8.1.0
rich>=13.7.0
python-dotenv>=1.0.0

# Azure SDK
azure-identity>=1.15.0
azure-mgmt-costmanagement>=4.0.0
azure-mgmt-resource>=23.0.0
azure-mgmt-subscription>=3.1.0

# AI for roasts using local Ollama
ollama>=0.1.0

# Data processing
pandas>=2.0.0
tabulate>=0.9.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
